requirements.txt
requests==2.27.1
tweepy==4.5.0
python-dotenv==0.19.2
